bump substreams client to sf.substreams.rpc.v2.Stream/Blocks
This breaks compatibility with previous substreams servers that only
support 'sf.substreams.v1.Stream/Blocks'
