+++ graph/build.rs
@@ -18,4 +18,11 @@ fn main() {
         .out_dir("src/substreams")
         .compile(&["proto/substreams.proto"], &["proto"])
         .expect("Failed to compile Substreams proto(s)");
+
+    tonic_build::configure()
+        .protoc_arg("--experimental_allow_proto3_optional")
+        .extern_path(".sf.substreams.v1", "crate::substreams")
+        .out_dir("src/substreams_rpc")
+        .compile(&["proto/substreams-rpc.proto"], &["proto"])
+        .expect("Failed to compile Substreams RPC proto(s)");
 }
