+++ graph/src/lib.rs
@@ -28,6 +28,8 @@ pub mod firehose;
 
 pub mod substreams;
 
+pub mod substreams_rpc;
+
 pub mod endpoint;
 
 /// Helpers for parsing environment variables.
