+++ graph/src/substreams_rpc/codec.rs
@@ -1,0 +1,5 @@
+#[rustfmt::skip]
+#[path = "sf.substreams.rpc.v2.rs"]
+mod pbsubstreamsrpc;
+
+pub use pbsubstreamsrpc::*;
