+++ graph/src/substreams_rpc/mod.rs
@@ -1,0 +1,3 @@
+mod codec;
+
+pub use codec::*;
